in (defn add-complement-to-bolt), (log/warn) rather than (exception) if you can’t generate a complement
It’s valid that expressions with a given specification can’t be
satisfied by any expression if the bolt is also supplying constraints
based on the parent phrase in the bolt. For example, we are generating
with spec1, and the complement parent phrase chosen randomly causes the
addition of spec s2 to the generation. unify(spec1,spec2) cannot be
generated, but there is another complement parent phrase  adding spec3,
where unify(spec1,spec2) can be generated.